ERA-13090: Refactor SchemaForm to normalize date-time separators, and to fix the timezone when injecting the value instead of updating the form data.
- Replaced `getFormDataWithFixedTimezones` with `normalizeDateTimeFieldValue` to handle date-time normalization directly in the form.
- Updated the rendering logic in `SchemaForm` to utilize the new normalization function for DATE_TIME fields.
- Removed the obsolete timezone correction utility and its associated tests.
